sentinel-main/pii_sentinel/requirements.txt
Flask>=3.0.0
pandas>=2.0.0
pdfminer.six>=20221105
python-docx>=1.0.0
openpyxl>=3.1.0

# File format parsers
xlrd>=2.0.1           # Legacy .xls Excel files
python-pptx>=0.6.21   # PowerPoint .pptx files
striprtf>=0.0.26      # RTF rich-text files
extract-msg>=0.48.0   # Outlook .msg email files

# IMAP email scanning (HTML body parsing — optional, falls back to stdlib)
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Cloud storage providers (install only those you use)
boto3>=1.34.0                      # AWS S3
google-api-python-client>=2.120.0  # Google Drive
google-auth>=2.28.0                # Google auth helper
azure-storage-blob>=12.19.0        # Azure Blob Storage
dropbox>=11.36.2                   # Dropbox

# MySQL database
mysql-connector-python>=8.3.0

# Local endpoint monitoring agent
watchdog>=4.0.0
